strategies moved to utils
